Make runpodcli dir user-specific
Also made it hidden, the user shouldn't need the folder, and I want to discourage users from putting stuff into it / changing it.

Note that multiple pods with the same settings overwrite the folder, but that's okay because the only pod-specific files are the ssh keys, and these are only used at startup.